Capitalize featuregate.Stage string values, remove Stage prefix
