+++ .chloggen/featuregatestage.yaml
@@ -1,0 +1,11 @@
+# One of 'breaking', 'deprecation', 'new_component', 'enhancement', 'bug_fix'
+change_type: breaking
+
+# The name of the component, or a single word describing the area of concern, (e.g. otlpreceiver)
+component: featuregate
+
+# A brief description of the change.  Surround your text with quotes ("") if it needs to start with a backtick (`).
+note: Capitalize `featuregate.Stage` string values, remove Stage prefix.
+
+# One or more tracking issues or pull requests related to the change
+issues: [6490]
