+++ featuregate/gates.go
@@ -19,39 +19,6 @@ import (
 	"sync"
 )
 
-// Stage represents the Gate's lifecycle and what is the expected state of it.
-type Stage int8
-
-const (
-	// StageAlpha is used when creating a new feature and the Gate must be explicitly enabled
-	// by the operator.
-	//
-	// The Gate will be disabled by default.
-	StageAlpha Stage = iota
-	// StageBeta is used when the feature flag is well tested and is enabled by default,
-	// but can be disabled by a Gate.
-	//
-	// The Gate will be enabled by default.
-	StageBeta
-	// StageStable is used when feature is permanently enabled and can not be disabled by a Gate.
-	// This value is used to provide feedback to the user that the gate will be removed in the next version.
-	//
-	// The Gate will be enabled by default and will return an error if modified.
-	StageStable
-)
-
-func (s Stage) String() string {
-	switch s {
-	case StageAlpha:
-		return "StageAlpha"
-	case StageBeta:
-		return "StageBeta"
-	case StageStable:
-		return "StageStable"
-	}
-	return "unknown"
-}
-
 // Gate is an immutable object that is owned by the `Registry`
 // to represents an individual feature that may be enabled or disabled based
 // on the lifecycle state of the feature and CLI flags specified by the user.
